NO-ISSUE: Reorganizing the repository
