+++ src/projects/prowjobsscraper/config.py
@@ -1,0 +1,4 @@
+import os
+
+JOB_LIST_URL = os.environ["JOB_LIST_URL"]
+LOG_LEVEL = os.getenv("LOG_LEVEL", "INFO")
